hashtags custom rendering working, pending 'onHashTagClick', wip
